[Refactor] Move BRPC stub cache to common
