+++ be/src/common/CMakeLists.txt
@@ -16,6 +16,9 @@ set(LIBRARY_OUTPUT_PATH "${BUILD_DIR}/src/common")
 
 set(COMMON_FILES
   bthread_timer.cpp
+  brpc/brpc_stub_cache.cpp
+  brpc/internal_service_recoverable_stub.cpp
+  brpc/lake_service_recoverable_stub.cpp
   brpc_helper.cpp
   configbase.cpp
   config.cpp
