+++ be/src/exec/data_sinks/data_stream_sender.cpp
@@ -48,6 +48,8 @@
 #include "base/compression/compression_utils.h"
 #include "base/uid_util.h"
 #include "column/chunk.h"
+#include "common/brpc/brpc_stub_cache.h"
+#include "common/brpc/internal_service_recoverable_stub.h"
 #include "common/config_compression_fwd.h"
 #include "common/config_exec_flow_fwd.h"
 #include "common/logging.h"
@@ -66,8 +68,6 @@
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "serde/protobuf_serde.h"
-#include "util/brpc_stub_cache.h"
-#include "util/internal_service_recoverable_stub.h"
 
 namespace starrocks {
