+++ be/src/exec/data_sinks/tablet_sink.cpp
@@ -50,6 +50,7 @@
 #include "column/map_column.h"
 #include "column/nullable_column.h"
 #include "column/raw_data_visitor.h"
+#include "common/brpc/brpc_stub_cache.h"
 #include "common/config_ingest_fwd.h"
 #include "common/config_scan_io_fwd.h"
 #include "common/statusor.h"
@@ -74,7 +75,6 @@
 #include "serde/protobuf_serde.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet_manager.h"
-#include "util/brpc_stub_cache.h"
 #include "util/stack_util.h"
 #include "util/thrift_rpc_helper.h"
