+++ be/src/exec/data_sinks/tablet_sink_index_channel.h
@@ -28,6 +28,7 @@
 #include "base/brpc/reusable_closure.h"
 #include "base/compression/block_compression.h"
 #include "base/container/raw_container.h"
+#include "common/brpc/internal_service_recoverable_stub.h"
 #include "common/status.h"
 #include "common/thread/threadpool.h"
 #include "common/tracer_fwd.h"
@@ -36,7 +37,6 @@
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "runtime/mem_tracker.h"
-#include "util/internal_service_recoverable_stub.h"
 
 namespace starrocks {
