+++ be/src/exec/pipeline/exchange/sink_buffer.h
@@ -26,11 +26,11 @@
 #include "base/brpc/disposable_closure.h"
 #include "base/phmap/phmap.h"
 #include "base/utility/defer_op.h"
+#include "common/brpc/internal_service_recoverable_stub.h"
 #include "exec/pipeline/fragment_context.h"
 #include "runtime/current_thread.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_fwd.h"
-#include "util/internal_service_recoverable_stub.h"
 
 namespace starrocks::pipeline {
