+++ be/src/runtime/exec_env.cpp
@@ -42,6 +42,7 @@
 #include "base/string/parse_util.h"
 #include "base/time/time.h"
 #include "base/utility/pretty_printer.h"
+#include "common/brpc/brpc_stub_cache.h"
 #include "common/config_exec_env_fwd.h"
 #include "common/config_lake_fwd.h"
 #include "common/config_vector_index_fwd.h"
@@ -114,7 +115,6 @@
 #include "storage/update_manager.h"
 #include "types/hll.h"
 #include "udf/python/env.h"
-#include "util/brpc_stub_cache.h"
 #include "util/priority_thread_pool.hpp"
 
 #ifdef STARROCKS_JIT_ENABLE
