+++ be/src/storage/segment_replicate_executor.cpp
@@ -20,6 +20,7 @@
 #include <utility>
 
 #include "base/container/raw_container.h"
+#include "common/brpc/brpc_stub_cache.h"
 #include "common/brpc_helper.h"
 #include "common/config_compaction_fwd.h"
 #include "fs/fs_posix.h"
@@ -29,7 +30,6 @@
 #include "runtime/load_fail_point.h"
 #include "runtime/mem_tracker.h"
 #include "storage/delta_writer.h"
-#include "util/brpc_stub_cache.h"
 
 namespace starrocks {
