+++ be/test/exec/pipeline/schedule/pipeline_timer_test.cpp
@@ -26,8 +26,8 @@
 
 #include "base/testutil/assert.h"
 #include "butil/time.h"
+#include "common/brpc/brpc_stub_cache.h"
 #include "gtest/gtest.h"
-#include "util/brpc_stub_cache.h"
 
 namespace starrocks::pipeline {
