+++ be/test/http/datacache_action_test.cpp
@@ -22,9 +22,9 @@
 #include "cache/data_cache_hit_rate_counter.hpp"
 #include "cache/disk_cache/starcache_engine.h"
 #include "cache/disk_cache/test_cache_utils.h"
+#include "common/brpc/brpc_stub_cache.h"
 #include "http/http_channel.h"
 #include "http/http_request.h"
-#include "util/brpc_stub_cache.h"
 
 namespace starrocks {
